+++ packages/adapter-components/src/config/shared.ts
@@ -15,7 +15,7 @@
 */
 import _ from 'lodash'
 import {
-  ElemID, ObjectType, BuiltinTypes, FieldDefinition, ListType, MapType,
+  ElemID, ObjectType, BuiltinTypes, FieldDefinition, MapType,
 } from '@salto-io/adapter-api'
 import { createMatchingObjectType } from '@salto-io/adapter-utils'
 import type { TransformationConfig, TransformationDefaultConfig } from './transformation'
@@ -45,7 +45,6 @@ export type AdapterApiConfig<
 
 export type UserFetchConfig = {
   includeTypes: string[]
-  hideTypes?: boolean
 }
 
 export const createAdapterApiConfigType = ({
@@ -103,19 +102,6 @@ export const createAdapterApiConfigType = ({
   })
   return adapterApiConfigType
 }
-
-export const createUserFetchConfigType = (
-  adapter: string,
-  additionalFields?: Record<string, FieldDefinition>,
-): ObjectType => (
-  new ObjectType({
-    elemID: new ElemID(adapter, 'userFetchConfig'),
-    fields: {
-      includeTypes: { refType: new ListType(BuiltinTypes.STRING) },
-      ...additionalFields,
-    },
-  })
-)
 
 export const getConfigWithDefault = <
   T extends TransformationConfig | FetchRequestConfig | undefined,
