+++ packages/adapter-components/test/config/ducktype.test.ts
@@ -14,7 +14,7 @@
 * limitations under the License.
 */
 import { ObjectType, BuiltinTypes, MapType } from '@salto-io/adapter-api'
-import { createDucktypeAdapterApiConfigType, createUserFetchConfigType, validateDuckTypeFetchConfig, validateDuckTypeApiDefinitionConfig } from '../../src/config'
+import { createDucktypeAdapterApiConfigType, createDuckTypeUserFetchConfigType, validateDuckTypeFetchConfig, validateDuckTypeApiDefinitionConfig } from '../../src/config'
 
 describe('config_ducktype', () => {
   describe('createAdapterApiConfigType', () => {
@@ -83,9 +83,10 @@ describe('config_ducktype', () => {
 
   describe('createUserFetchConfigType', () => {
     it('should return default type when no custom fields were added', () => {
-      const type = createUserFetchConfigType('myAdapter')
-      expect(Object.keys(type.fields)).toHaveLength(1)
+      const type = createDuckTypeUserFetchConfigType('myAdapter')
+      expect(Object.keys(type.fields)).toHaveLength(2)
       expect(type.fields.includeTypes).toBeDefined()
+      expect(type.fields.hideTypes).toBeDefined()
     })
   })
